Map between SomaScan SeqIds and Gene Ontology (GO) identifiers

SomaScanGO is an R object that provides mappings between SeqIds and the GO identifiers that they are directly associated with. This mapping and its reverse mapping (SomaScanGO2PROBE) do not associate the child terms from the GO ontology with the gene. Only the directly evidenced terms are represented here.

SomaScanGO2ALLPROBES is an R object that provides mappings between a given GO identifier and all of the manufacturer identifiers annotated at that GO term or to one of its child nodes in the GO ontology. Thus, this mapping is much larger and more inclusive than SomaScanGO2PROBE.

Details

If SomaScanGO is cast as a list, each SeqId is mapped to a list of lists. The names of the outer list are GO identifiers. Each inner list consists of three named elements: GOID, Ontology, and Evidence.

The GOID element matches the GO identifier named in the outer list and is included for convenience when processing the data using lapply().

The Ontology element indicates which of the three Gene Ontology categories this identifier belongs to. The categories are biological process (BP), cellular component (CC), and molecular function (MF).

The Evidence element contains a code indicating what kind of evidence supports the association of the GO identifier to the SeqId. Some of the evidence codes in use include:

IMP: inferred from mutant phenotype

IGI: inferred from genetic interaction

IPI: inferred from physical interaction

ISS: inferred from sequence similarity

IDA: inferred from direct assay

IEP: inferred from expression pattern

IEA: inferred from electronic annotation

TAS: traceable author statement

NAS: non-traceable author statement

ND: no biological data available

IC: inferred by curator

A more complete listing of evidence codes can be found at http://www.geneontology.org/GO.evidence.shtml

If SomaScanGO2ALLPROBES or SomaScanGO2PROBE is cast as a list, each GO term maps to a named vector of SeqIds and evidence codes. A GO identifier may be mapped to the same manufacturer identifier more than once but the evidence code can be different. Mappings between Gene Ontology identifiers, Gene Ontology terms, and other information are available in a separate data package named GO.

Whenever any of these mappings are cast as a data.frame, all the results will be output in an appropriate tabular form.

Mappings between SeqIds and GO information were obtained through their mappings to SeqIds. NA valuess are assigned to SeqIds that cannot be mapped to any Gene Ontology information.

All mappings were based on data provided by Gene Ontology (http://current.geneontology.org/ontology/go-basic.obo), with a date stamp from the source of: 2021-09-01

Value

A Bimap object of the ProbeAnnDbBimap class.

References

ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/

See also

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
x <- SomaScanGO
# Get the manufacturer identifiers that are mapped to a GO ID
mapped_genes <- mappedkeys(x)
# Convert to a list
xx <- as.list(x[mapped_genes][1:300])
if(length(xx) > 0) {
    # Try the first one
    got <- xx[[1]]           
    got[[1]][["GOID"]]
    got[[1]][["Ontology"]]
    got[[1]][["Evidence"]]
}
#> [1] "IBA"
# For the reverse map:
x <- SomaScanGO2PROBE
mapped_genes <- mappedkeys(x)
# Convert to a list
xx <- as.list(x[mapped_genes][1:300])
if(length(xx) > 0){
    # Gets the manufacturer ids for the top 2nd and 3nd GO identifiers
    goids <- xx[2:3]
    # Gets the manufacturer ids for the first element of goids
    goids[[1]]
    # Evidence code for the mappings
    names(goids[[1]])
}
#>   [1] "IBA" "IBA" "IMP" "IEA" "IEA" "IMP" "IMP" "IEA" "IEA" "IBA" "IMP"
#>  [12] "IEA" "IEA" "IBA" "IEA" "TAS" "IMP" "IEA" "TAS" "IEA" "IEA" "IMP"
#>  [23] "IEA" "IEA" "IBA" "IBA" "IMP" "IEA" "IMP" "IBA" "IDA" "TAS" "IEA"
#>  [34] "IEA" "TAS" "TAS" "TAS" "IMP" "TAS" "IEA" "IEA" "TAS" "IBA" "IEA"
#>  [45] "IBA" "TAS" "IBA" "TAS" "IEA" "IBA" "NAS" "IEA" "TAS" "IBA" "ISS"
#>  [56] "TAS" "IEA" "IEA" "IEA" "TAS" "TAS" "IEA" "TAS" "IEA" "TAS" "IEA"
#>  [67] "IEA" "ISS" "TAS" "IBA" "IMP" "IEA" "TAS" "TAS" "TAS" "TAS" "TAS"
#>  [78] "TAS" "IEA" "IEA" "IEA" "TAS" "IEA" "IEA" "TAS" "IEA" "IEA" "IEA"
#>  [89] "IEA" "TAS" "IEA" "IEA" "IEA" "TAS" "IBA" "IMP" "IEA" "IMP" "IEA"
#> [100] "TAS" "IEA" "IEA" "IEA" "IMP" "TAS" "IMP" "IEA" "TAS" "IEA" "IEA"
#> [111] "IEA" "IEA" "TAS"

x <- SomaScanGO2ALLPROBES
mapped_genes <- mappedkeys(x)
# Convert SomaScanGO2ALLPROBES to a list
xx <- as.list(x[mapped_genes][1:300])
if(length(xx) > 0){
# Gets the manufacturer identifiers for the top 2nd and 3nd GO identifiers
    goids <- xx[2:3]
    # Gets all the manufacturer identifiers for the first element of goids
    goids[[1]]
    # Evidence code for the mappings
    names(goids[[1]])
}
#>  [1] "IBA" "IBA" "IEA" "IEA" "IBA" "ISS" "IEA" "IEA" "IEA" "IEA" "ISS"
#> [12] "IEA" "IEA" "IEA" "IEA" "IEA" "IEA" "IEA" "IBA" "IBA" "IBA" "IEA"
#> [23] "IEA" "ISS" "IEA" "ISS" "IEA" "IBA" "IBA" "IEA" "IBA" "ISS" "IEA"
#> [34] "ISS" "IEA" "ISS" "IDA" "ISS" "ISS" "IEA" "ISS" "IEA" "IEA" "IEA"
#> [45] "ISS" "IEA" "ISS" "ISS" "IEA" "IEA" "ISS" "IEA" "ISS" "ISS" "IEA"
#> [56] "IEA" "IEA" "ISS" "ISS" "IEA" "IEA" "IEA" "IEA" "IEA" "IEA"