Add Target Full Name information to data frame

Retrieve the Target Full Name for a given SeqId (or set of SeqIds) and add that information to a data frame. The Target Full Name can be found in an ADAT file and represents the full name of the protein target of a primary SOMAmer reagent.

Note: You can retrieve a list of Target Full Name values for a given set of SeqIds with mget(seqids, SomaScanTARGETFULLNAME), where seqids is a character vector of SeqIds. To retrieve a data frame of all Target Full Name values present in the package, use toTable(SomaScanTARGETFULLNAME).

Usage

addTargetFullName(data)

Arguments

data

a data frame generated by a select query. This data frame can have any number of columns, but must contain a PROBEID column. This will be used to match the Target Full Name.

Value

The input data frame with one additional column called "TARGETFULLNAME".

Examples

seqs <- withr::with_seed(123, sample(keys(SomaScan.db), 5))
res <- select(SomaScan.db, keys = seqs, columns = "UNIPROT")
#> 'select()' returned 1:many mapping between keys and columns
addTargetFullName(res)
#>     PROBEID
#> 1   18943-4
#> 2   18943-4
#> 3   18943-4
#> 4   18943-4
#> 5   18943-4
#> 6   18943-4
#> 7  19168-71
#> 8  19168-71
#> 9  19168-71
#> 10 19168-71
#> 11 19168-71
#> 12 19168-71
#> 13 19168-71
#> 14 19168-71
#> 15 20998-19
#> 16 20998-19
#> 17 20998-19
#> 18 20998-19
#> 19  4842-62
#> 20  4842-62
#> 21  4842-62
#> 22  4842-62
#> 23  4842-62
#> 24  4842-62
#> 25  4842-62
#> 26  4842-62
#> 27  4842-62
#> 28  4842-62
#> 29   9287-6
#> 30   9287-6
#> 31   9287-6
#> 32   9287-6
#> 33   9287-6
#>                                                             TARGETFULLNAME
#> 1                                                Prostaglandin reductase 2
#> 2                                                Prostaglandin reductase 2
#> 3                                                Prostaglandin reductase 2
#> 4                                                Prostaglandin reductase 2
#> 5                                                Prostaglandin reductase 2
#> 6                                                Prostaglandin reductase 2
#> 7               Proline-serine-threonine phosphatase-interacting protein 1
#> 8               Proline-serine-threonine phosphatase-interacting protein 1
#> 9               Proline-serine-threonine phosphatase-interacting protein 1
#> 10              Proline-serine-threonine phosphatase-interacting protein 1
#> 11              Proline-serine-threonine phosphatase-interacting protein 1
#> 12              Proline-serine-threonine phosphatase-interacting protein 1
#> 13              Proline-serine-threonine phosphatase-interacting protein 1
#> 14              Proline-serine-threonine phosphatase-interacting protein 1
#> 15 Nicotinamide riboside kinase 2; Short=NRK 2; Short=NmR-K 2; EC=2.7.1.22
#> 16 Nicotinamide riboside kinase 2; Short=NRK 2; Short=NmR-K 2; EC=2.7.1.22
#> 17 Nicotinamide riboside kinase 2; Short=NRK 2; Short=NmR-K 2; EC=2.7.1.22
#> 18 Nicotinamide riboside kinase 2; Short=NRK 2; Short=NmR-K 2; EC=2.7.1.22
#> 19                                                              Glypican-3
#> 20                                                              Glypican-3
#> 21                                                              Glypican-3
#> 22                                                              Glypican-3
#> 23                                                              Glypican-3
#> 24                                                              Glypican-3
#> 25                                                              Glypican-3
#> 26                                                              Glypican-3
#> 27                                                              Glypican-3
#> 28                                                              Glypican-3
#> 29                                                   Sepiapterin reductase
#> 30                                                   Sepiapterin reductase
#> 31                                                   Sepiapterin reductase
#> 32                                                   Sepiapterin reductase
#> 33                                                   Sepiapterin reductase
#>       UNIPROT
#> 1      Q3L8A4
#> 2      Q6MZH8
#> 3      Q8N8N7
#> 4      V9HW32
#> 5      G3V3Y1
#> 6      G3V2R9
#> 7      O43586
#> 8      J3KPG6
#> 9  A0A0S2Z5P3
#> 10     B5BU74
#> 11     B5BUK4
#> 12     O43585
#> 13     O95657
#> 14     H0YKY3
#> 15     B7ZKR3
#> 16     Q52M81
#> 17     Q9NZK3
#> 18     Q9NPI5
#> 19 A0A1S5UZ08
#> 20     Q53H15
#> 21     Q8IYG2
#> 22     B4DTD8
#> 23     P51654
#> 24     C9JLE3
#> 25     G3V1R0
#> 26     I6QTG3
#> 27     Q2L880
#> 28     Q2L882
#> 29     A8K741
#> 30     Q53GI9
#> 31     Q9UBB1
#> 32     D6W5H2
#> 33     P35270